Properties: robustness of comparison with _notHasKey constant
In _Lookup class, the value of hasKey should not be stored at
the time of instantiation, because in the frequent case where it is
_notHasKey, by the time it is compared with _notHasKey, the latter may
have changed.

This is not so exotic as it sounds: it happens easily if a reload occurs
between instantiation and comparison, for instance. On two of my buildmasters
(0.8.10 and 0.8.12), this started happening for another, unknwon reason