Quickfix for DirectoryUpload hanging if lacking permissions
The difference with the correct (from end-user perspective)
behaviour with FileUpload lies in the fact that with the latter,
the remote commands for upload don't even start, because
reservation of the tempfile fails.

Tracking down what happens with the error during unpack has
not been very succesful for me so far, it's quite intricated.
But it's preoccupying : if it failed for another reason (FS full, say),
it would still hang, and I suspect FileUpload would behave in a similar
way. This will need lost of simulation to assess.

Anyway it seems a good idea also to maximize the chances that the
temp file lies on the same FS as the destination, so this quickfix also
fixes another bug.